+++ kamehouse-shell/my.scripts/win/video-playlists/create-all-video-playlists-http-media-server.sh
@@ -13,10 +13,10 @@ if [ "$?" != "0" ]; then
 fi
 
 PATH_PLS_SOURCE=${HOME}/git/texts/video_playlists/linux
-PATH_PLS_DEST=${HOME}/git/texts/video_playlists/http-niko-server
+PATH_PLS_DEST=${HOME}/git/texts/video_playlists/http-media-server
 
 PATH_BASE_SOURCE="/media/media-drive"
-PATH_BASE_DEST="http://niko-server/kame-house-streaming/media-server/media-drive"
+PATH_BASE_DEST="http://${MEDIA_SERVER}/kame-house-streaming/media-server/media-drive"
 
 mainProcess() {
   validateVariables